Player switching control schemes works as expected.
