+++ Assets/RiderFlow.UserData/Editor/Scenes/Bookmarks.Map/f933bc58b1ea17142aaa4b15307f4adf.userdata
@@ -1,0 +1,7 @@
+{
+  "GlobalObjectId_V1-2-f933bc58b1ea17142aaa4b15307f4adf-2131096962-0": {
+    "GameObjectId": "GlobalObjectId_V1-2-f933bc58b1ea17142aaa4b15307f4adf-2131096962-0",
+    "AssetGUID": "f933bc58b1ea17142aaa4b15307f4adf",
+    "UniqueItemID": "GlobalObjectId_V1-2-f933bc58b1ea17142aaa4b15307f4adf-2131096962-0"
+  }
+}
