+++ Assets/RiderFlow.UserData/Editor/Scenes/Sections.Map/2cda990e2423bbf4892e6590ba056729.userdata
@@ -11,22 +11,22 @@
   },
   "GlobalObjectId_V1-2-2cda990e2423bbf4892e6590ba056729-1271871563-0": {
     "GameObjectId": "GlobalObjectId_V1-2-2cda990e2423bbf4892e6590ba056729-1271871563-0",
-    "SectionColorRed": 0.219607845,
-    "SectionColorGreen": 0.219607845,
-    "SectionColorBlue": 0.219607845,
-    "SectionAltColorRed": 0.219607845,
-    "SectionAltColorGreen": 0.219607845,
-    "SectionAltColorBlue": 0.219607845,
+    "SectionColorRed": 0.172549024,
+    "SectionColorGreen": 0.172549024,
+    "SectionColorBlue": 0.172549024,
+    "SectionAltColorRed": 0.172549024,
+    "SectionAltColorGreen": 0.172549024,
+    "SectionAltColorBlue": 0.172549024,
     "RemoveSectionDuringBuild": false
   },
   "GlobalObjectId_V1-2-2cda990e2423bbf4892e6590ba056729-1185005051-0": {
     "GameObjectId": "GlobalObjectId_V1-2-2cda990e2423bbf4892e6590ba056729-1185005051-0",
-    "SectionColorRed": 0.219607845,
-    "SectionColorGreen": 0.219607845,
-    "SectionColorBlue": 0.219607845,
-    "SectionAltColorRed": 0.219607845,
-    "SectionAltColorGreen": 0.219607845,
-    "SectionAltColorBlue": 0.219607845,
+    "SectionColorRed": 0.172549024,
+    "SectionColorGreen": 0.172549024,
+    "SectionColorBlue": 0.172549024,
+    "SectionAltColorRed": 0.172549024,
+    "SectionAltColorGreen": 0.172549024,
+    "SectionAltColorBlue": 0.172549024,
     "RemoveSectionDuringBuild": false
   },
   "GlobalObjectId_V1-2-2cda990e2423bbf4892e6590ba056729-305330804-0": {
@@ -61,32 +61,32 @@
   },
   "GlobalObjectId_V1-2-2cda990e2423bbf4892e6590ba056729-324567166-0": {
     "GameObjectId": "GlobalObjectId_V1-2-2cda990e2423bbf4892e6590ba056729-324567166-0",
-    "SectionColorRed": 0.219607845,
-    "SectionColorGreen": 0.219607845,
-    "SectionColorBlue": 0.219607845,
-    "SectionAltColorRed": 0.219607845,
-    "SectionAltColorGreen": 0.219607845,
-    "SectionAltColorBlue": 0.219607845,
+    "SectionColorRed": 0.172549024,
+    "SectionColorGreen": 0.172549024,
+    "SectionColorBlue": 0.172549024,
+    "SectionAltColorRed": 0.172549024,
+    "SectionAltColorGreen": 0.172549024,
+    "SectionAltColorBlue": 0.172549024,
     "RemoveSectionDuringBuild": false
   },
   "GlobalObjectId_V1-2-2cda990e2423bbf4892e6590ba056729-1967245234-0": {
     "GameObjectId": "GlobalObjectId_V1-2-2cda990e2423bbf4892e6590ba056729-1967245234-0",
-    "SectionColorRed": 0.219607845,
-    "SectionColorGreen": 0.219607845,
-    "SectionColorBlue": 0.219607845,
-    "SectionAltColorRed": 0.219607845,
-    "SectionAltColorGreen": 0.219607845,
-    "SectionAltColorBlue": 0.219607845,
+    "SectionColorRed": 0.172549024,
+    "SectionColorGreen": 0.172549024,
+    "SectionColorBlue": 0.172549024,
+    "SectionAltColorRed": 0.172549024,
+    "SectionAltColorGreen": 0.172549024,
+    "SectionAltColorBlue": 0.172549024,
     "RemoveSectionDuringBuild": false
   },
   "GlobalObjectId_V1-2-2cda990e2423bbf4892e6590ba056729-1902932414-0": {
     "GameObjectId": "GlobalObjectId_V1-2-2cda990e2423bbf4892e6590ba056729-1902932414-0",
-    "SectionColorRed": 0.219607845,
-    "SectionColorGreen": 0.219607845,
-    "SectionColorBlue": 0.219607845,
-    "SectionAltColorRed": 0.219607845,
-    "SectionAltColorGreen": 0.219607845,
-    "SectionAltColorBlue": 0.219607845,
+    "SectionColorRed": 0.172549024,
+    "SectionColorGreen": 0.172549024,
+    "SectionColorBlue": 0.172549024,
+    "SectionAltColorRed": 0.172549024,
+    "SectionAltColorGreen": 0.172549024,
+    "SectionAltColorBlue": 0.172549024,
     "RemoveSectionDuringBuild": false
   },
   "GlobalObjectId_V1-2-2cda990e2423bbf4892e6590ba056729-1701742105-0": {
@@ -97,6 +97,16 @@
     "SectionAltColorRed": 0.219607845,
     "SectionAltColorGreen": 0.219607845,
     "SectionAltColorBlue": 0.219607845,
+    "RemoveSectionDuringBuild": false
+  },
+  "GlobalObjectId_V1-2-2cda990e2423bbf4892e6590ba056729-901159214-0": {
+    "GameObjectId": "GlobalObjectId_V1-2-2cda990e2423bbf4892e6590ba056729-901159214-0",
+    "SectionColorRed": 0.172549024,
+    "SectionColorGreen": 0.172549024,
+    "SectionColorBlue": 0.172549024,
+    "SectionAltColorRed": 0.172549024,
+    "SectionAltColorGreen": 0.172549024,
+    "SectionAltColorBlue": 0.172549024,
     "RemoveSectionDuringBuild": false
   }
 }
