+++ Assets/RiderFlow.UserData/Editor/Scenes/Sections.Map/f504a935c6bbfbf45aaaf8d220905e4c.userdata
@@ -1,0 +1,12 @@
+{
+  "GlobalObjectId_V1-2-f504a935c6bbfbf45aaaf8d220905e4c-12733306-0": {
+    "GameObjectId": "GlobalObjectId_V1-2-f504a935c6bbfbf45aaaf8d220905e4c-12733306-0",
+    "SectionColorRed": 0.172549024,
+    "SectionColorGreen": 0.172549024,
+    "SectionColorBlue": 0.172549024,
+    "SectionAltColorRed": 0.172549024,
+    "SectionAltColorGreen": 0.172549024,
+    "SectionAltColorBlue": 0.172549024,
+    "RemoveSectionDuringBuild": false
+  }
+}
