+++ Assets/_Project/Runtime/PREFABS/PlayerConfiguration.prefab
@@ -11,6 +11,7 @@ GameObject:
   - component: {fileID: 7711285624575338378}
   - component: {fileID: 7972469277145718601}
   - component: {fileID: 4062431423150835742}
+  - component: {fileID: -7907636821074277119}
   - component: {fileID: 703169452674830338}
   - component: {fileID: 825156129046926009}
   m_Layer: 0
@@ -19,7 +20,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 0
+  m_IsActive: 1
 --- !u!4 &7711285624575338378
 Transform:
   m_ObjectHideFlags: 0
@@ -140,6 +141,18 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   playerInput: {fileID: 7972469277145718601}
   localPlayerID: 0
+--- !u!114 &-7907636821074277119
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 785687840274375511}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 1a1e41b60a895e54abcebe171310786f, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!114 &703169452674830338
 MonoBehaviour:
   m_ObjectHideFlags: 0
