+++ Assets/_Project/Runtime/Scenes/Scrapped_CharacterSelect.unity.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f933bc58b1ea17142aaa4b15307f4adf
+guid: 382ed846ac001d841b70ef83daf1a7bd
 DefaultImporter:
   externalObjects: {}
   userData:
