+++ Assets/_Project/Runtime/Scripts/Utility/PlayerUINavigationManager.cs
@@ -1,0 +1,4 @@
+public class PlayerUINavigationManager : SingletonPersistent<PlayerUINavigationManager>
+{
+    
+}
