+++ Assets/_Project/Runtime/Scripts/Utility/PlayerUINavigationManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8ed733a50875d5d4c8a5b0455cd592c6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
