+++ ProjectSettings/EditorBuildSettings.asset
@@ -6,8 +6,11 @@ EditorBuildSettings:
   serializedVersion: 2
   m_Scenes:
   - enabled: 1
+    path: Assets/_Project/Runtime/Scenes/Dev_MainMenu.unity
+    guid: f504a935c6bbfbf45aaaf8d220905e4c
+  - enabled: 1
     path: Assets/_Project/Runtime/Scenes/Dev_CharacterSelect.unity
-    guid: 382ed846ac001d841b70ef83daf1a7bd
+    guid: f933bc58b1ea17142aaa4b15307f4adf
   - enabled: 1
     path: Assets/_Project/Runtime/Scenes/Dev_Game.unity
     guid: 2cda990e2423bbf4892e6590ba056729
